Move hook to src/features/datasets
